+++ src/ll/misc.py
@@ -18,8 +18,91 @@ LivingLogic modules and packages.
 import sys, types, collections, weakref
 
 
-# fetch item, first, last and count
-from ll._misc import *
+# fetch item, first, last, count and xmlescape
+try:
+	from ll._misc import *
+except ImportError:
+	# Use Python implementations of those functions
+	_defaultitem = object()
+
+	def item(iterator, index, default=_defaultitem):
+		"""
+		Returns the :var:`index`'th element from the iterable. :var:`index` may be
+		negative to count from the end. E.g. 0 returns the first element produced
+		by the iterator, 1 the second, -1 the last one etc. If :var:`index` is
+		negative the iterator will be completely exhausted, if it's positive it
+		will be exhausted up to the :var:`index`'th element. If the iterator
+		doesn't produce that many elements :exc:`IndexError` will be raised,
+		except when :var:`default` is given, in which case :var:`default` will be
+		returned.
+		"""
+		i = index
+		if i>=0:
+			for item in iterator:
+				if not i:
+					return item
+				i -= 1
+		else:
+			i = -index
+			cache = collections.deque()
+			for item in iterator:
+				cache.append(item)
+				if len(cache)>i:
+					cache.popleft()
+			if len(cache)==i:
+				return cache.popleft()
+		if default is _defaultitem:
+			raise IndexError(index)
+		else:
+			return default
+
+
+	def first(iterator, default=_defaultitem):
+		"""
+		Return the first element from the iterable. If the iterator doesn't
+		produce any elements :exc:`IndexError` will be raised, except when
+		:var:`default` is given, in which case :var:`default` will be returned.
+		"""
+		return item(iterator, 0, default)
+
+
+	def last(iterator, default=_defaultitem):
+		"""
+		Return the last element from the iterable. If the iterator doesn't produce
+		any elements :exc:`IndexError` will be raised, except when :var:`default`
+		is given, in which case :var:`default` will be returned.
+		"""
+		return item(iterator, -1, default)
+
+
+	def count(iterator):
+		"""
+		Count the number of elements produced by the iterable. Calling this
+		function will exhaust the iterator.
+		"""
+		count = 0
+		for node in iterator:
+			count += 1
+		return count
+
+	def xmlescape(string):
+		"""
+		Return a copy of the argument string, where every occurrence of ``<``,
+		``>``, ``&``, ``\"``, ``'`` and every restricted character has been
+		replaced with their XML character entity or character reference.
+		"""
+		maps = (
+			("&", "&amp;"),
+			("<", "&lt;"),
+			(">", "&gt;"),
+			("'", "&apos;"),
+			('"', "&quot;"),
+		)
+		for (s, r) in maps:
+			string = string.replace(s, r)
+		for c in "\x00\x01\x02\x03\x04\x05\x06\x07\x08\x0b\x0c\x0e\x0f\x10\x11\x12\x13\x14\x15\x16\x17\x18\x19\x1a\x1b\x1c\x1d\x1f\x7f\x80\x81\x82\x83\x84\x86\x87\x88\x89\x8a\x8b\x8c\x8d\x8e\x8f\x90\x91\x92\x93\x94\x95\x96\x97\x98\x99\x9a\x9b\x9c\x9d\x9e\x9f":
+			string = string.replace(c, "&#%d;" % ord(c))
+		return string
 
 
 __docformat__ = "reStructuredText"