+++ test/test_xist_detox.py
@@ -144,13 +144,4 @@ def test_textexpr():
 			+detox.code("""s = '"a" < "b" & "b" > "a"'""")
 			+detox.textexpr("s")
 
-	assert makeoutput(e, "gurk") == '"a" &lt; "b" &amp; "b" &gt; "a"'
-
-
-def test_attrexpr():
-	with xsc.Frag() as e:
-		with defblock(func="gurk()"):
-			+detox.code("""s = '"a" < "b" & "b" > "a"'""")
-			+detox.attrexpr("s")
-
 	assert makeoutput(e, "gurk") == '&quot;a&quot; &lt; &quot;b&quot; &amp; &quot;b&quot; &gt; &quot;a&quot;'